feat: load real Flask package when installed
